storage: introduce read-only type
Fixes #461

We introduce the read-only variant of Storage. The use of this variant
implies that it will be used for read-only purposes, and so it does not
hold a signer.

Storage is re-formulated to be a combination of the read-only store
along with the signer and fetchers.
